Rename file, remove obsolete DEBUG_COLORS code
